Struct bittrex::Client

pub struct Client { /* fields omitted */ }

Bittrex API client.

Methods

impl Client

Public API methods.

pub fn get_markets(&self) -> Result<Box<[MarketInfo]>>

Used to get the open and available trading markets at Bittrex along with other meta data.

pub fn get_currencies(&self) -> Result<Box<[CurrencyInfo]>>

Used to get all supported currencies at Bittrex along with other meta data.

pub fn get_ticker<S: AsRef<str>>(&self, market: S) -> Result<TickerInfo>

Used to get the current tick values for a market.

pub fn get_market_summaries(&self) -> Result<Box<[MarketSummary]>>

Used to get the last 24 hour summary of all active exchanges.

pub fn get_market_summary<S: AsRef<str>>(
&self,
market: S
) -> Result<MarketSummary>

Used to get the last 24 hour summary of the given market.

pub fn get_order_book<S: AsRef<str>>(
&self,
market: S,
order_type: OrderBookType,
depth: u8
) -> Result<OrderBook>

Used to get retrieve the order book for a given market.

Panics if the depth is bigger than 50.

pub fn get_market_history<S: AsRef<str>>(
&self,
market: S
) -> Result<MarketSummary>

Used to retrieve the latest trades that have occured for a specific market.

impl Client

Private API methods.

Note: All this methods will panic if the client is not logged in.

pub fn get_balances(&self) -> Result<Box<[BalanceInfo]>>

Gets the balances of the Bittrex account.

Note: it will panic if not logged in.

impl Client

pub fn new() -> Result<Client>

Creates a new client with default configuration.

pub fn login<K: Into<String>, S: Into<String>>(
&mut self,
api_key: K,
api_secret: S
)

Logs the client in for future private method access.

Note: This will not perform any request to Bittrex, it will only store the API key and the secret for future use.