Struct bittrex::types::CurrencyInfo

pub struct CurrencyInfo { /* fields omitted */ }

Currency information structure.

Methods

impl CurrencyInfo

pub fn currency(&self) -> &str

Gets the 3-letter currency code.

pub fn currency_long(&self) -> &str

Gets the long currency name.

pub fn min_confirmation(&self) -> u32

Gets the minimum number of confirmations to credit the account.

pub fn tx_fee(&self) -> f32

Gets the transaction fee.

pub fn is_active(&self) -> bool

Gets wether the currency is active or not.

pub fn coin_type(&self) -> &str

Gets the coin type string constant.

pub fn base_address(&self) -> Option<&String>

Gets the base address for the coin at Bittrex.

pub fn notice(&self) -> Option<&String>

Gets the optional notice about the currency.