Struct bittrex::types::MarketSummary

pub struct MarketSummary { /* fields omitted */ }

Market summary structure

Methods

impl MarketSummary

pub fn market_name(&self) -> &str

Gets the name of the market.

pub fn high(&self) -> Option<f32>

Gets the highest transaction value in the last 24 hours for the market.

pub fn low(&self) -> Option<f32>

Gets the lowest transaction value in the last 24 hours for the market.

pub fn last(&self) -> Option<f32>

Gets the last transaction price.

pub fn bid(&self) -> Option<f32>

Gets the current bidding/buying price.

pub fn ask(&self) -> Option<f32>

Gets the current asking/selling price.

pub fn volume(&self) -> Option<f32>

Gets the volume of the market.

pub fn base_volume(&self) -> Option<f32>

Gets the base volume of the market.

pub fn time_stamp(&self) -> NaiveDateTime

Gets the timestamp of the information.

pub fn open_buy_orders(&self) -> Option<u32>

Gets the number of open buying orders.

pub fn open_sell_orders(&self) -> Option<u32>

Gets the number of open selling orders.

pub fn prev_day(&self) -> f32

Gets the price of the previous day.

pub fn created(&self) -> NaiveDateTime

Gets the market creation time.

pub fn display_market_name(&self) -> Option<&String>

Gets the name to display for the market.