Struct bittrex::Client

pub struct Client {
    inner: Client,
    api_key: Option<String>,
    api_secret: Option<String>,
}

Bittrex API client.

Fields
inner: Client

Inner reqwest client.

api_key: Option<String>

API key.

api_secret: Option<String>

API secret.

Methods

impl Client

Private API methods.

Note: All this methods will panic if the client is not logged in.

pub fn get_balances(&self) -> Result<Vec<BalanceInfo>, Error>

Gets the balances of the Bittrex account.

Note: it will panic if not logged in.

impl Client

Public API methods.

pub fn get_markets(&self) -> Result<Vec<MarketInfo>, Error>

Used to get the open and available trading markets at Bittrex along with other meta data.

pub fn get_currencies(&self) -> Result<Vec<CurrencyInfo>, Error>

Used to get all supported currencies at Bittrex along with other meta data.

pub fn get_ticker<S: AsRef<str>>(&self, market: S) -> Result<TickerInfo, Error>

Used to get the current tick values for a market.

pub fn get_market_summaries(&self) -> Result<Vec<MarketSummary>, Error>

Used to get the last 24 hour summary of all active exchanges.

pub fn get_market_summary<S: AsRef<str>>(
&self,
market: S
) -> Result<MarketSummary, Error>

Used to get the last 24 hour summary of the given market.

pub fn get_order_book<S: AsRef<str>>(
&self,
market: S,
order_type: OrderBookType,
depth: u8
) -> Result<OrderBook, Error>

Used to get retrieve the order book for a given market.

Panics if the depth is bigger than 50.

pub fn get_market_history<S: AsRef<str>>(
&self,
market: S
) -> Result<MarketSummary, Error>

Used to retrieve the latest trades that have occurred for a specific market.

impl Client

pub fn login<K: Into<String>, S: Into<String>>(
&mut self,
api_key: K,
api_secret: S
)

Logs the client in for future private method access.

Note: This will not perform any request to Bittrex, it will only store the API key and the secret for future use.

fn append_login(&self, url: &mut Url)

Appends the login information to the query string.

Note: it will panic if not logged in.

fn get_headers(&self, url: &Url) -> Result<HeaderMap, Error>

Gets the headers for the given URL.

Note: it will panic if not logged in.

fn hash_uri(&self, url: &Url) -> Result<String, Error>

Hash the given URI with the logged in API secret.

Note: it will panic if not logged in.