Struct bittrex::types::MarketSummary

pub struct MarketSummary {
    market_name: String,
    high: Option<f32>,
    low: Option<f32>,
    last: Option<f32>,
    bid: Option<f32>,
    ask: Option<f32>,
    volume: Option<f32>,
    base_volume: Option<f32>,
    time_stamp: NaiveDateTime,
    open_buy_orders: Option<u32>,
    open_sell_orders: Option<u32>,
    prev_day: f32,
    created: NaiveDateTime,
    display_market_name: Option<String>,
}

Market summary structure

Fields
market_name: String

Name of the market.

high: Option<f32>

Highest transaction value in the last 24 hours for the market.

low: Option<f32>

Lowest transaction value in the last 24 hours for the market.

last: Option<f32>

Last transaction price.

bid: Option<f32>

Current bidding/buying price.

ask: Option<f32>

Current asking/selling price.

volume: Option<f32>

Volume of the market.

base_volume: Option<f32>

Base volume of the market.

time_stamp: NaiveDateTime

Timestamp of the information.

open_buy_orders: Option<u32>

Number of open buying orders.

open_sell_orders: Option<u32>

Number of open selling orders.

prev_day: f32 created: NaiveDateTime

Market creation time.

display_market_name: Option<String>

Name to display for the market.

Methods

impl MarketSummary

pub fn market_name(&self) -> &str

Gets the name of the market.

pub fn high(&self) -> Option<f32>

Gets the highest transaction value in the last 24 hours for the market.

pub fn low(&self) -> Option<f32>

Gets the lowest transaction value in the last 24 hours for the market.

pub fn last(&self) -> Option<f32>

Gets the last transaction price.

pub fn bid(&self) -> Option<f32>

Gets the current bidding/buying price.

pub fn ask(&self) -> Option<f32>

Gets the current asking/selling price.

pub fn volume(&self) -> Option<f32>

Gets the volume of the market.

pub fn base_volume(&self) -> Option<f32>

Gets the base volume of the market.

pub fn time_stamp(&self) -> NaiveDateTime

Gets the timestamp of the information.

pub fn open_buy_orders(&self) -> Option<u32>

Gets the number of open buying orders.

pub fn open_sell_orders(&self) -> Option<u32>

Gets the number of open selling orders.

pub fn prev_day(&self) -> f32

Gets the price of the previous day.

pub fn created(&self) -> NaiveDateTime

Gets the market creation time.

pub fn display_market_name(&self) -> Option<&String>

Gets the name to display for the market.